Develop

Un pò di migliorie alla gestione delle assenze e gestione dell'inserimento assenze fuori sede dei dipendenti.

See merge request !202

+++ app/manager/services/absences/model/Scanner.java
@@ -167,8 +167,12 @@ public class Scanner {
       }
       //eseguire
       for (AbsenceTrouble toDelete : toDeleteTroubles) {
-        log.info("Rimuovo problem {} {}", absence.toString(), toDelete.trouble);
-        toDelete.delete();
+        if (toDelete.isPersistent()) { //FIXME Issue #324
+          toDelete.refresh();
+          log.info("Rimuovo problem {} {}", absence.toString(), toDelete.trouble);
+          toDelete.delete();  
+        }
+        
       }
       for (AbsenceTrouble toAdd : toAddTroubles) {
         log.info("Aggiungo problem {} {}", absence.toString(), toAdd.trouble);
